# api/openapi.yaml
openapi: 3.1.0
info:
  title: FlowScore API
  version: 0.1.0
  description: |
    Match page slice. Source of truth for routes: design/mockup/ApiMap.dc.html.
    Flow semantics: docs/FLOW.md.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
tags:
  - name: matches

paths:
  /matches:
    get:
      tags: [matches]
      operationId: listMatches
      summary: Every match for the home screen, by kick-off time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchList'
        default:
          $ref: '#/components/responses/Error'


  /matches/{matchId}:
    get:
      tags: [matches]
      operationId: getMatch
      summary: Match header, score and stats
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: OK
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '304':
          description: Not modified (If-None-Match matched)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'

  /matches/{matchId}/flow:
    get:
      tags: [matches]
      operationId: getMatchFlow
      summary: Current Flow and per-minute history
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowSeries'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'

  /matches/{matchId}/events:
    get:
      tags: [matches]
      operationId: getMatchEvents
      summary: Match timeline, newest first
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'

  /matches/{matchId}/insight:
    get:
      tags: [matches]
      operationId: getMatchInsight
      summary: AI output — explanation, outcome prediction, fan sentiment
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Insight'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'

  /matches/{matchId}/simulate:
    post:
      tags: [matches]
      operationId: simulateMatch
      summary: '"What if?": the outcome chances with changes later in the match'
      description: |
        Predicts the rest of the match from the current moment, once as it
        stands and once with the changes (docs/PREDICTION.md). Nothing is stored.
      parameters:
        - $ref: '#/components/parameters/MatchId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Simulation'
        '400':
          description: A change the model cannot take
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'

  /ws/matches/{matchId}/stream:
    servers:
      - url: ws://localhost:8080
    get:
      tags: [matches]
      operationId: streamMatch
      summary: WebSocket — live updates for one match
      description: |
        Upgrades to WebSocket. The server pushes WsMessage frames (JSON).
        On a gap in `seq` the client refetches the REST resources.
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '101':
          description: Switching protocols. Frames follow the WsMessage schema.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WsMessage'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    MatchId:
      name: matchId
      in: path
      required: true
      schema:
        type: string
      example: m_01J8ZQ4T7K

  headers:
    ETag:
      schema:
        type: string

  responses:
    Unauthorized:
      description: Access token missing or expired
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Match not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Error:
      description: Unexpected error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

  schemas:
    Problem:
      description: RFC 9457 Problem Details
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
          description: Shown in the error card ("req 7f3a91")

    Side:
      type: string
      enum: [home, away]

    Team:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: string
        code:
          type: string
          description: Three-letter code, e.g. MCI
        name:
          type: string

    Score:
      type: object
      required: [home, away]
      properties:
        home:
          type: integer
          minimum: 0
        away:
          type: integer
          minimum: 0

    Match:
      type: object
      required: [id, status, kickoffAt, competition, home, away, score, clock, stats, seq]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [scheduled, live, halftime, finished, postponed]
        kickoffAt:
          type: string
          format: date-time
        competition:
          type: object
          required: [id, name]
          properties:
            id:
              type: string
            name:
              type: string
            round:
              type: string
        venue:
          type: object
          properties:
            name:
              type: string
            city:
              type: string
        home:
          $ref: '#/components/schemas/Team'
        away:
          $ref: '#/components/schemas/Team'
        score:
          $ref: '#/components/schemas/Score'
        halftimeScore:
          oneOf:
            - $ref: '#/components/schemas/Score'
            - type: 'null'
        clock:
          $ref: '#/components/schemas/Clock'
        stats:
          type: array
          items:
            $ref: '#/components/schemas/StatRow'
        seq: 
          $ref : '#/components/schemas/SnapshotSeq'


    MatchList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MatchSummary'

    MatchSummary:
      description: |
        One card of the home screen: the match plus what the card draws. The
        list is polled, so seq here does not count as a stream position.
      allOf:
        - $ref: '#/components/schemas/Match'
        - type: object
          required: [flow, delta10, factors, points, prediction, explanation]
          properties:
            flow:
              $ref: '#/components/schemas/FlowValues'
            delta10:
              $ref: '#/components/schemas/FlowValues'
            factors:
              type: array
              items:
                $ref: '#/components/schemas/FlowFactor'
            points:
              type: array
              items:
                $ref: '#/components/schemas/FlowPoint'
            prediction:
              oneOf:
                - $ref: '#/components/schemas/Prediction'
                - type: 'null'
            explanation:
              oneOf:
                - $ref: '#/components/schemas/Explanation'
                - type: 'null'





    Clock:
      type: object
      required: [elapsedSeconds, period, observedAt]
      properties:
        elapsedSeconds:
          type: integer
          description: Match time in seconds (72:14 → 4334)
        period:
          type: string
          enum: [first_half, second_half, extra_time, penalties]
        observedAt:
          type: string
          format: date-time
          description: Server time when elapsedSeconds was measured; the client keeps ticking from here

    StatRow:
      type: object
      required: [key, home, away]
      properties:
        key:
          type: string
          description: >-
            possession (percent), shots, shots_on_target, xg, key_passes, pressure_tackles
            (tackles won or lost), in the order of the Match board
        home:
          type: number
        away:
          type: number

    FlowValues:
      type: object
      required: [home, away]
      properties:
        home:
          type: number
          minimum: 0
          maximum: 100
        away:
          type: number
          minimum: 0
          maximum: 100

    FlowPoint:
      type: object
      required: [minute, home, away]
      properties:
        minute:
          type: integer
          minimum: 1
        home:
          type: number
        away:
          type: number

    FlowSeries:
      type: object
      required: [matchId, current, delta10, points, factors, updatedAt, seq]
      properties:
        matchId:
          type: string
        current:
          $ref: '#/components/schemas/FlowValues'
        delta10:
          type: object
          description: Flow now minus Flow 10 match minutes ago
          required: [home, away]
          properties:
            home:
              type: number
            away:
              type: number
        points:
          type: array
          description: One point per played minute. The client derives the wave and 1/5/15 windows.
          items:
            $ref: '#/components/schemas/FlowPoint'
        factors:
          type: array
          items:
            $ref: '#/components/schemas/FlowFactor'
        updatedAt:
          type: string
          format: date-time
        seq:
          $ref: '#/components/schemas/SnapshotSeq'

    FlowFactor:
      type: object
      description: >-
        What one group of events adds to a team's impulse right now (docs/FLOW.md,
        section 6). The factors of a team add up to its impulse, so they explain
        its Flow exactly; groups under half a point are left out.
      required: [side, key, value, count, minutes]
      properties:
        side:
          $ref: '#/components/schemas/Side'
        key:
          type: string
          enum: [shots, goals, key_passes, possession, corners, cards, red_cards, substitutions, other]
          description: cards are the opponent's yellow cards, red_cards the team's own
        value:
          type: number
          description: Impulse points, the scale of the Flow-эффект in the chronicle; negative when it holds the team back
        count:
          type: integer
          description: Events of the group in the last 10 minutes of the period; for possession, minutes
        minutes:
          type: integer
          description: Minutes since the first of them, rounded up; 0 when count is 0
        share:
          type: number
          minimum: 0
          maximum: 1
          description: Possession only, the team's share of the ball over those minutes

    MatchEvent:
      type: object
      required: [id, type, side, minute]
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            [goal, shot_on_target, shot_off_target, shot_blocked, corner,
             yellow_card, red_card, substitution, foul, offside, take_on,
             halftime, fulltime]
        side:
          oneOf:
            - $ref: '#/components/schemas/Side'
            - type: 'null'
        minute:
          type: integer
          description: The minute on the clock, as the live badge shows it (72:14 → 72)
        addedTime:
          type: [integer, 'null']
          description: 45+2 → minute 45, addedTime 2
        player:
          $ref: '#/components/schemas/PersonRef'
        assist:
          $ref: '#/components/schemas/PersonRef'
        xG:
          type: [number, 'null']
        position:
          type: object
          description: 0–100, attacking left to right
          properties:
            x:
              type: number
            y:
              type: number
        scoreAfter:
          $ref: '#/components/schemas/Score'
        flowImpact:
          type: [number, 'null']
          description: The event's weight w(e) from docs/FLOW.md

    PersonRef:
      type: [object, 'null']
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string

    EventList:
      type: object
      required: [items, seq]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MatchEvent'
        seq:
          $ref: '#/components/schemas/SnapshotSeq'

    SnapshotSeq:
      type: integer
      description: |
        Seq of the last stream message already included in this response.
        Apply only stream messages with a greater seq.


    Probabilities:
      type: object
      description: Whole percents of a home win, a draw and an away win; they add up to 100
      required: [home, draw, away]
      properties:
        home:
          type: integer
        draw:
          type: integer
        away:
          type: integer

    Insight:
      type: object
      required: [matchId, prediction, seq]
      properties:
        matchId:
          type: string
        explanation:
          description: Null until the Explainability agent answers
          oneOf:
            - $ref: '#/components/schemas/Explanation'
            - type: 'null'
        prediction:
          $ref: '#/components/schemas/Prediction'
        sentiment:
          oneOf:
            - $ref: '#/components/schemas/Sentiment'
            - type: 'null'
        seq:
          $ref: '#/components/schemas/SnapshotSeq'

    Prediction:
      type: object
      description: |
        Chances of the result from the outcome model (docs/PREDICTION.md).
        current changes with prediction.update; preMatch is fixed at kick-off.
      required: [current, preMatch, model]
      properties:
        current:
          $ref: '#/components/schemas/Probabilities'
        preMatch:
          $ref: '#/components/schemas/Probabilities'
        model:
          type: string
          example: Poisson v1

    SimulationRequest:
      type: object
      required: [reds]
      properties:
        reds:
          type: array
          maxItems: 8
          items:
            $ref: '#/components/schemas/SimulatedRed'

    SimulatedRed:
      type: object
      description: A red card later in the match; a minute already past counts from now
      required: [side, minute]
      properties:
        side:
          $ref: '#/components/schemas/Side'
        minute:
          type: integer
          minimum: 1
          maximum: 120

    Simulation:
      type: object
      required: [minute, current, scenario]
      properties:
        minute:
          type: integer
          description: Match minute on the clock the simulation starts from
        current:
          $ref: '#/components/schemas/Probabilities'
        scenario:
          $ref: '#/components/schemas/Probabilities'

    Explanation:
      type: object
      description: |
        The analysis the AI agent wrote for one moment of the match: a goal, a
        red card or every tenth minute. The factors on the card come from the
        flow, so the analysis itself is text only.
      required: [title, text, minute, generatedAt]
      properties:
        title:
          type: string
          example: Арсенал забирает инициативу
        text:
          type: string
        minute:
          type: integer
          description: The match minute the analysis was written for
        generatedAt:
          type: string
          format: date-time


    Sentiment:
      type: object
      required: [messagesPerMinute, home, away]
      properties:
        messagesPerMinute:
          type: integer
        home:
          $ref: '#/components/schemas/FanMood'
        away:
          $ref: '#/components/schemas/FanMood'

    FanMood:
      type: object
      required: [score, mood]
      properties:
        score:
          type: number
          minimum: -1
          maximum: 1
        mood:
          type: string
          example: эйфория

    WsMessage:
      oneOf:
        - $ref: '#/components/schemas/WsMatchEvent'
        - $ref: '#/components/schemas/WsFlowUpdate'
        - $ref: '#/components/schemas/WsMatchStats'
        - $ref: '#/components/schemas/WsPredictionUpdate'
        - $ref: '#/components/schemas/WsInsightUpdate'
      discriminator:
        propertyName: type
        mapping:
          match.event: '#/components/schemas/WsMatchEvent'
          flow.update: '#/components/schemas/WsFlowUpdate'
          match.stats: '#/components/schemas/WsMatchStats'
          prediction.update: '#/components/schemas/WsPredictionUpdate'
          insight.update: '#/components/schemas/WsInsightUpdate'

    WsEnvelope:
      type: object
      required: [type, matchId, seq, sentAt]
      properties:
        type:
          type: string
        matchId:
          type: string
        seq:
          type: integer
          description: Increases by 1 per message within a match; a gap means messages were lost
        sentAt:
          type: string
          format: date-time

    WsMatchEvent:
      allOf:
        - $ref: '#/components/schemas/WsEnvelope'
        - type: object
          required: [data]
          properties:
            type:
              const: match.event
            data:
              $ref: '#/components/schemas/MatchEvent'

    WsFlowUpdate:
      allOf:
        - $ref: '#/components/schemas/WsEnvelope'
        - type: object
          required: [data]
          properties:
            type:
              const: flow.update
            data:
              type: object
              required: [current, delta10, point, clock, factors]
              properties:
                current:
                  $ref: '#/components/schemas/FlowValues'
                delta10:
                  $ref: '#/components/schemas/FlowValues'
                point:
                  $ref: '#/components/schemas/FlowPoint'
                clock:
                  $ref: '#/components/schemas/Clock'
                  description: Match time at this update; the client resyncs its ticking clock to it
                factors: 
                  type: array
                  items:
                    $ref: '#/components/schemas/FlowFactor'


    WsMatchStats:
      allOf:
        - $ref: '#/components/schemas/WsEnvelope'
        - type: object
          required: [data]
          properties:
            type:
              const: match.stats
            data:
              type: object
              required: [stats]
              description: Every row, not a diff; the client replaces what it has
              properties:
                stats:
                  type: array
                  items:
                    $ref: '#/components/schemas/StatRow'

    WsPredictionUpdate:
      allOf:
        - $ref: '#/components/schemas/WsEnvelope'
        - type: object
          required: [data]
          properties:
            type:
              const: prediction.update
            data:
              $ref: '#/components/schemas/Probabilities'

    WsInsightUpdate:
      allOf:
        - $ref: '#/components/schemas/WsEnvelope'
        - type: object
          required: [data]
          properties:
            type:
              const: insight.update
            data:
              $ref: '#/components/schemas/Explanation'